Move the loader cache to separate files
